begin implementation of SidFileUtils class

+++ lib/sidfileutils.h
@@ -1,6 +1,8 @@
 #ifndef SidFileUtils_h
 #define SidFileUtils_h
 
+#include <cstdio>
+
 #include "constants.h"
 #include "sidexception.h"
 
@@ -8,7 +10,7 @@ class SidFileUtils {
 
   public:
     static const bool exists(const char *filename);
-    static const byte *read(const char *filename, unsigned int &filesize);
+    static const byte *read(const char *filename, unsigned long int &filesize);
     static void write(const char *filename, const unsigned int filesize, const byte *data);
 };
